requirements.txt
streamlit==1.41.1
weaviate-client==4.10.2
sentence-transformers==3.3.1